Refactored old javabeans inside SeatSelectionActivity
